Print all elements of config_vars_to_export
Before, this would only print the first config var, even if you
specified multiple ones.

It still exported them all, it just only told you it was exporting the
first one.

"Referring to the content of a member variable of an array without
providing an index number is the same as referring to the content of the
first element, the one referenced with index number zero."

(from:
http://tldp.org/LDP/Bash-Beginners-Guide/html/sect_10_02.html#sect_10_02_02)