Allow additional test arguments as config vars
Right now you can't do things like add a coverage report or specify
which folders to constrain the testing to. This additional config var,
when present, will simply be added to the end of `mix test`. If it is
not present, or the value is an empty string, then `mix test` will
execute normally (i.e. as though no arguments were passed to it).