use MIX_HOME,HEX_HOME when installing mix/hex dependencies
tl;dr:

* Instead of relying on Mix's default behavior of installing to
`$HOME/.mix`, the buildpack sets `$MIX_HOME` to a directory controlled
by the buildpack when running `mix`
* Instead of relying on Hex's default behavior of installing to
`$HOME/.hex`, the buildpack sets `$HEX_HOME` to a directory controlled
by the buildpack when running `hex`
* The buildpack now exports `HEX_HOME` and `MIX_HOME` if they are not
set as configuration variables.

By using `MIX_HOME` and `HEX_HOME`, we can consolidate some
logic for transferring mix/hex files in and out of the cache and app build
directory.

This should allow the buildpack to work in the new build system and the
old build system since this buildpack no longer relies on what the value
of `HOME` is at either build time or runtime.

fixes #194

https://github.com/HashNuke/heroku-buildpack-elixir/issues/194

The Longer Explanation
========================

This strategy was taken from the [heroku-buildpack-ruby buildpack][1], which sets
[GEM_HOME and GEM_PATH environment variables][2] as well as the [`BUNDLE_PATH` environment variable][3]
to a path in the `BUILD_DIR` argument when running the `bundle install` command.

Although the Elixir programming langauge and buildpack doesn't use
Bundler/Rubygems, it has similar patterns for setting the installation
directory and archive paths for Mix and Hex.

`MIX_HOME` has been supported since `v0.7.0` of Elixir ([support for `MIX_HOME` added ~9 years ago][4]).

Note: There is a `MIX_ARCHIVES` env variable that Mix can use, but [MIX_ARCHIVES support was added
in version 0.14.7][5]. I'm not sure if anyone is still using this Elixir version
(they should upgrade as the OTP that 0.14.7 supports is very outdated and
 doesn't receive security updates). If `MIX_ARCHIVES` is not present, then [Mix
will put archives in `$MIX_HOME/archives`][6]. In order to keep compatibility as close as possible,
this commit continues using the old strategy of backing up everything in `MIX_HOME`, rather than what's in `MIX_ARCHIVES`
alone.

[`HEX_HOME` support was added in Hex v0.3.3][7]. This was the commit that made
Hex put stuff in `$HEX_HOME` directory, meaning that support for `HEX_HOME` has
existed for a long time and we should be able to rely on Hex using the
`HEX_HOME` directory.  The old strategy used by the Elixir buildpack always
backed up the default `HEX_HOME` value (`$HOME/.hex`).

[1]: https://github.com/heroku/heroku-buildpack-ruby
[2]: https://github.com/heroku/heroku-buildpack-ruby/blob/52e02d56370c1e05434bc550a06f1bcde5849644/lib/language_pack/ruby.rb#L324-L325
[3]: https://github.com/heroku/heroku-buildpack-ruby/blob/52e02d56370c1e05434bc550a06f1bcde5849644/lib/language_pack/ruby.rb#L349
[4]: https://github.com/elixir-lang/elixir/commit/e518c5263e66d6554c7229d170a94b8afac25f70#diff-ff4ab450307d771fedb0471df59d9bd71f4f90559375d57f8015fda8ba4cf301R21
[5]: https://github.com/elixir-lang/elixir/commit/02a415facdfc86373eebaf507b83061b291fd6d2#diff-75787cbe3255513265f665bb859b99a624db2b56bec01e48455cd63577ad8cadR16
[6]: https://github.com/elixir-lang/elixir/blob/c57d08a32f1308769a2fbaf7a2fe385ab2fb5d22/lib/mix/lib/mix.ex#L463
[7]: https://github.com/hexpm/hex/commit/5dd8ae020dc4c31bc662e490c79221423fc5d9f6